dags/datalake/fetch_open_brewery/requirements.txt
apache-airflow==2.10.2
apache-airflow-providers-amazon==9.0.0
boto3==1.35.47
deltalake==0.20.2
logging==0.4.9.6
pandas==2.2.3
requests==2.32.3